PR: Removed Generic BoxResource
- within this realese we will use "offset"'s based pagging instead of "next_mark"
